refactor: remove enshrined v7_0_0 upgrade in op-deployer (#766)
